Test_Automation/Project_Recipe_Blog/Test_Automation/requirements.txt
selenium
pytest
allure-pytest
allure-python-commons
requests
mysql-connector-python
